+++ packages/react/lib/main.tsx
@@ -14,6 +14,8 @@ export type * from "types";
 export { Reveal } from "./ui/Reveal";
 export { Card } from "./ui/Card";
 export { Pin } from "./ui/Pin";
+export { ThreeDSecure } from "./ui/ThreeDSecure";
+export { useThreeDSecure } from "./ui/useThreeDSecure";
 export { useEvervault, themes };
 
 export interface CustomConfig extends BrowserConfig {