+++ e2e-apps/decrypt-backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @evervault/e2e-decrypt-backend
 
+## 1.0.3
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- @evervault/browser-e2e-app@1.0.1
+
 ## null
 
 ### Patch Changes