+++ packages/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @evervault/react
 
+## 2.9.0
+
+### Minor Changes
+
+- 3156568: Add autoProgress option to Card component to automatically progress to the next input when an input becomes valid.
+
+## 2.8.0
+
+### Minor Changes
+
+- 1df9ba2: Add functionality to disable autoCompletion for fields in the Card component
+
+## 2.7.0
+
+### Minor Changes
+
+- 05eda8e: Adds new ThreeDSecure UI Component for handling 3D Secure authentication
+
 ## 2.6.1
 
 ### Patch Changes